Separate out the engine that runs the scheduler and the workers and the client that allows jobs to be scheduled and canceled.
